max77xxx: Make sure that the RTC is a wakeup interrupt
The max77xxx uses IRQ domains to distribute interrupts to the
different parts of the MFD.  That means we have a separate IRQ number
(a virtual IRQ) that is in charge of the RTC.

Up until now we've marked the main max77xxx interrupt as a "wakeup"
interrupt, but not the virtual one.  Typically it doesn't matter a
whole lot, but there's a tiny window during the suspend process where
it's relevant.  Specifically, I believe that the order is:

1. The RTC interrupt goes off right as we're running
   suspend_device_irqs().  We kick off a task to start the thread that
   handles the max77xxx interrupt.

2. We quickly finish suspending the device_irqs(), including the
   virutal IRQ for the max77xxx RTC.

3. When the max77xxx thread runs and calls handle_nested_irq() finally
   it adds it as a pending interrupt.

4. check_wakeup_irqs() will properly cancel the suspend only if the
   RTC interrupt itself was marked as a wakeup interrupt.

We now properly set the RTC as a wakeup interrupt.

BUG=chrome-os-partner:27224
TEST=suspend_stress_test
TEST=Patch in Ifd370b79689bf13de359d640a83bbf8c5b5bed52 and see that
it prevents suspend.

Change-Id: I6baa5a260513471e91e25cebaf84f31d942f0f77
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191962
Reviewed-by: Olof Johansson <olofj@chromium.org>