max77xxx: Use irq_set_parent; don't claim to be an edge interrupt
The irq_set_chip_and_handler() takes a function that's supposed to be
the interrupt handler, but this function is in practice never really
called in the case of IRQ Domains.

It's probably a good thing that it's never really called because we
were passing a totally bogus function.  The handle_edge_irq() will
unconditionally call our ack() function which we define as NULL.

I say "never really" because there's one place where the function is
called.  ...and that's causing a crash.  The function is called if we
don't have a "parent" and we're resending a deffered IRQ at resume
time (see resend_irqs()).

Fix the resume time crash in two ways:

* Use irq_set_parent() which will make the resume-time code call the
  parent handler, not the one passed to irq_set_chip_and_handler().
  This now makes it so that the function passed is _actually_ never
  called.  The irq_set_parent() was added specifically to address this
  case.

* Pass handle_simple_irq() instead of handle_edge_irq().  This is more
  benign so if it's called we'll get a warning like:
    Primary handler called for nested irq X

BUG=chrome-os-partner:27224
TEST=suspend_stress_test
TEST=Patch in Ifd370b79689bf13de359d640a83bbf8c5b5bed52 and no longer
see crashes.

Change-Id: Ied036db9d9426bfaea6ee47edde8ca79ae698aac
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191961
Reviewed-by: Olof Johansson <olofj@chromium.org>